+++ samples/petclinic/src/org/springframework/samples/petclinic/web/FindOwnersForm.java
@@ -39,8 +39,9 @@ public class FindOwnersForm	extends AbstractClinicForm {
 	}
 
 	/** Method used to search for owners renders View depending on how many are found */
-	protected ModelAndView onSubmit(HttpServletRequest request, HttpServletResponse response,
-																	Object command, BindException errors) throws Exception {
+	protected ModelAndView onSubmit(
+			HttpServletRequest request, HttpServletResponse response, Object command, BindException errors)
+			throws Exception {
 
 		Owner owner = (Owner) command;
 
@@ -60,7 +61,7 @@ public class FindOwnersForm	extends AbstractClinicForm {
 
 		// 1 owner found
 		owner = (Owner) results.iterator().next();
-		return new ModelAndView(getSuccessView(), "ownerId", Long.toString(owner.getId()));
+		return new ModelAndView(getSuccessView(), "ownerId", Integer.toString(owner.getId()));
 	}
 
 }
